Path point A and B now work with a left click
Ganked dynamic light stuff

MobileIsValid() uses the path mobile to adjust the position to valid if possible, or reject it.  It bumps the box up until it is free, then drops it to the ground.

MobileCanReach() sets the footing for movement and then tries a few move iterations to reach the desired point