PathGraph skips impassible connections when returning them for draw
When finding connections, only allow upward movement in a stairstep amount, this prevents pathing leaping up onto things.  Jump pathing needs to be a separate step.

Can set nodes to blocked via picking in TestPathing

In CanReach(), return if point a is not on ground.  Not too sure about this one