A little contents enum for Quake1 as that's what I'm testing at the moment
Lots of adjustments to player movement physics and the addition of swimming.  Swimming still needs climb out functionality.

Break out movement into functions

Double map space to jump on press and release (probably not a good idea), and climb / swim up / fly up on press-and-hold

Dive and crouch are in the actions but I haven't done those yet