Using new entitylib for handling moving models, staticy entities etc
Only draw / update entities in the PVS.

Entity updates are super super slow.  I didn't really think there were enough in a typical quake map to slow anything down (300ish?), but it is brutal.  Need to profile and see if some of them are doing something crazy or if it is just the usual OOP overhead.